move some handlers, fix test
